+++ CHANGELOG.md
@@ -7,8 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.0.6] / 2023-11-24
 ### Features
 - `Tests` project.
+### PackageBuilder
+- Enable `DocumentationFile` in package.
+- Update `BuilderUtils` to return instance.
 ### Tests
 - Test `Constants`
+- Test `BuilderUtils`
 
 ## [1.0.5] / 2021-12-21
 - Update Build Project